+++ CHANGELOG.md
@@ -64,6 +64,7 @@ These are only breaking changes for unformatted code.
 - Fix issue where `foo(x,_)` in uncurried mode would generate a curried function https://github.com/rescript-lang/rescript-compiler/pull/6082
 - Fix printing of uncurried application when the lhs is a function definition https://github.com/rescript-lang/rescript-compiler/pull/6084
 - Fix parsing uncurried type starting with path https://github.com/rescript-lang/rescript-compiler/pull/6089
+- Fix bigInt comparison https://github.com/rescript-lang/rescript-compiler/pull/6097
 
 #### :nail_care: Polish